Create a Command Line Interface (CLI) for the enigma-simulator.
The enigma-simulator implementation was intended for use with Jupyter
Notebook and a command line interface was not provided.  Having a CLI
may help to provide broader and more immediate use of the
enigma-simulator project since Jupyter Notebook is not required.

The commit for this CLI consists of a single file, enigma_cli.py,
which implements a CLI and simply imports and uses the pre-existing
"machine" implementation.  Only the CLI file has been added, no other
project files have been touched.

Here's the CLI usage message:

usage: enigma_cli.py [-h] [-w wheel_order] [-m message_key]
                     [-s plugboard_pairs] [-p]

This Python script is provided as a convenience to run the
enigma-simulator Python implementation from the command line.  The
original project was intended to be run in a Jupyter Notebook
environment.

optional arguments:
  -h, --help          show this help message and exit
  -w wheel_order      the 3-digit wheel order from left to right, default is
                      123; each digit represents the corresponding enigma
                      wheel number, supported wheel values are 1, 2, 3, and 5
  -m message_key      the initial 3-letter wheel setting from left wheel to
                      right wheel; default is AAA
  -s plugboard_pairs  single string consisting of one to six pairs of letters
                      to swap through the plugboard (steckerbrett), default is
                      no pairs swapped; for example to swap A & Z and Y & B
                      use "-s AZYB"
  -p                  prompt user for text input, default is to read directly
                      from standard input (stdin)

Note that the enigma support provided allows you to select the wheels,
their order & initial letter position and also set plugboard values.
Under the covers a B reflector is used and the wheel rings are AAA.
Output is delivered to standard out (stdout).

Example usage from the command line, first using the optional "-p"
argument to prompt the user for input and the second delivering the
input text via stdin:
    $ ./enigma_cli.py -w 312 -m XYZ -s EDCTAB -p
    Enter your text and press <Enter>: ENIGMACRYPTO
    AFMCXZJQIVZA
    $ echo AFMCXZJQIVZA | ./enigma_cli.py -w 312 -m XYZ -s EDCTAB
    ENIGMACRYPTO
    $